requirements.txt
selenium
webdriver-manager
beautifulsoup4
requests
discord.py
pyqt5
pyinstaller 